add game logic with players default names
