+++ index.html
@@ -11,10 +11,19 @@
   <body>
     
     <div id="input" class="">
-      <label for="name" class="player-input">Player One</label>
-      <input type="text" id="name1">
-      <label for="name" class="player-input">Player Two</label>
-      <input type="text" id="name2">
+      <h1>Hi welcome to tic tac toe game!</h1>
+      <div class="" id="player-1">
+        <p>Enter Player One's name below or press next to use default names</p>
+        <label for="name" class="player-input">Player One</label>
+        <input type="text" id="name1" placeholder="Player X">
+        <button id="player1-next">Next</button>
+      </div>
+      <div class="" id="player-2">
+        <p>Enter Player Two's name below or press next to use default names</p>
+        <label for="name" class="player-input">Player Two</label>
+        <input type="text" id="name2" placeholder="Player O">
+        <button id="player1-next">Next</button>
+      </div>
       <button id="add-players">Add Player</button>
     </div>
