+++ js/game.js
@@ -1,0 +1,9 @@
+const gameBoard = (() => {
+  const board = ['X', 'O', 'X', 'O', 'X', 'O', 'X', 'O', 'X'];
+})();
+
+const Player = name => { name };
+
+
+
+export { gameBoard, Player };
